feat: Update CountryList component with responsive height
The code changes in the `CountryList` component adjust the height of the component based on the number of countries being displayed. If there are less than or equal to 4 countries, the height is set to `auto`. Otherwise, the height is set to `90vh` to ensure the component takes up most of the viewport. This change improves the responsiveness of the `CountryList` component.